Update avatar background color and typewriter options
Changed the background color of the avatar to #50cef4 and updated the typewriter options in the Home/Type.js file. The avatar now has a blue background and the typewriter options include "Software Engineer" and "Full Stack Engineer". This commit improves the visual appearance and messaging on the home page.